+++ js/data.js
@@ -1,24 +1,71 @@
-import {getNumberPhoto} from '/js/util.js';
-import {getIdPhoto} from '/js/util.js';
-import {getUrlPhoto} from '/js/util.js';
-import {getDescription} from '/js/util.js';
-import {getRandomInteger} from '/js/util.js';
-import {createComment} from '/js/util.js';
+const NAMES = [
+  'Ольга',
+  'Иван',
+  'Елена',
+  'Анна',
+  'Сергей',
+  'Антон',
+  'Александр',
+  'Полина',
+  'Егор',
+  'Майя',
+  'Кирилл',
+  'Наталья',
+  'Илья',
+  'Маргарита',
+  'Макар',
+  'Ника',
+  'Федор',
+  'Ульяна',
+  'Алексей',
+  'Вера',
+  'Василий',
+  'Виктория',
+  'Валерий',
+  'София',
+  'Герман',
+  'Екатерина',
+  'Игорь',
+  'Марина',
+  'Тимофей',
+  'Ирина'
+];
 
+const DESCRIPTION_LIST = [
+  'Территория отеля',
+  'Указатель как дойти до пляжа',
+  'Океан',
+  'Девушка',
+  'Суп',
+  'Машина',
+  'Клубника на тарелке',
+  'Компот в стаканах',
+  'Самолет пролетает над пляжем',
+  'Обувница',
+  'Путь к океану',
+  'Машина белая ауди',
+  'Овощи и красная рыба',
+  'Котик-ролл',
+  'Теплые дутики',
+  'Самолет в небе',
+  'Выступление оркестра',
+  'Ретроавтомобиль',
+  'Тапочки с фонариками',
+  'Пальмы',
+  'Салат',
+  'Закат на берегу океана',
+  'Краб',
+  'Концерт',
+  'Джип в озере с бегемотом, крокодилом'
+];
 
-const createPhotoDescription = () => {
-  const numberPhoto = getNumberPhoto();
-  return {
-    id: getIdPhoto(),
-    url: getUrlPhoto(numberPhoto),
-    description: getDescription(numberPhoto),
-    likes: getRandomInteger(15, 200),
-    comments: Array.from({length: getRandomInteger(0, 30)}, createComment),
-  };
-};
+const MESSAGE_LIST = [
+  'Всё отлично! ',
+  'В целом всё неплохо. Но не всё. ',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально. ',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше. ',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше. ',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?! '
+];
 
-const PhotosDescription = () => Array.from({length: 25}, createPhotoDescription);
-
-export {PhotosDescription};
-// console.log(PhotosDescription);
-//JSON.stringify(PhotosDescription);
+export {NAMES, DESCRIPTION_LIST, MESSAGE_LIST};
